+++ modules/infotext.py
@@ -68,7 +68,10 @@ def parse(infotext):
     debug(f'Raw: {infotext}')
 
     remaining = infotext.replace('\nSteps:', ' Steps:')
-    params = ['steps:', 'seed:', 'width:', 'height:', 'sampler:', 'size:', 'cfg scale:'] # first param is one of those
+    # TODO infotext: handle using regex instead
+    params = [' steps:', ' seed:', ' width:', ' height:', ' sampler:', ' size:', ' cfg scale:'] # first param is one of those
+    params += ['\nsteps:', '\nseed:', '\nwidth:', '\nheight:', '\nsampler:', '\nsize:', '\ncfg scale:']
+    params += ['.steps:', '.seed:', '.width:', '.height:', '.sampler:', '.size:', '.cfg scale:']
 
     prompt_end = [remaining.lower().find(p) for p in params if p in remaining.lower()]
     prompt_end += [remaining.lower().find('negative prompt:')]