+++ CHANGELOG.md
@@ -25,6 +25,7 @@
   - zluda hotfixes  
   - zluda force sync  
   - fix torch import on compile  
+  - infotext parser force delimiter before params  
 
 ## Update for 2025-02-18